refactor/KAN-19 created test util file and moved renderwithwrapper function
