+++ src/pages/resetPassword/__tests__/ResetPassword.test.tsx
@@ -4,8 +4,8 @@ import { TEST_ID } from "../../../components/inputs/__tests__/testIds";
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { submitBtnTestId } from "../../../components/authForm/AuthForm";
-import { renderWithWrapper } from "./utils";
 import ResetPassword from "../ResetPassword";
+import { renderWithWrapper } from "../../../utils/tests";
 
 vi.mock("@tanstack/react-query", async () => {
   const actual = await vi.importActual("@tanstack/react-query");
